Add files via upload
For Allison: It works but not well, can you polish it up. The code is supposed to have a button that increases the counting number by 1. The code works in that there is a button and that it counts up but visually it looks really bad. Can you center everything up and fix the title?